Rename v3 file
