+++ packages/react-native-gesture-handler/src/v3/gestureStateManager.web.ts
@@ -2,7 +2,7 @@ import { tagMessage } from '../utils';
 import type IGestureHandler from '../web/handlers/IGestureHandler';
 import GestureHandlerOrchestrator from '../web/tools/GestureHandlerOrchestrator';
 import NodeManager from '../web/tools/NodeManager';
-import type { GestureStateManagerType } from './types/GestureStateManager';
+import type { GestureStateManagerType } from './types/GestureStateManagerTypes';
 
 function ensureHandlerAttached(handler: IGestureHandler) {
   if (!handler.attached) {
